Updated 10-07-18

+++ resources/views/admin/orders.blade.php
@@ -1,4 +1,4 @@
-<div class="container">
+<div class="">
     @section('center-column')
     	<h2>Список заказов</h2>
         <a href="{{ route('admin.order.create') }}" class="btn btn-primary">Добавить заказ</a>
@@ -22,28 +22,23 @@
             </tr>
             </thead>
 			<tbody>
+               
     			@forelse($orders as $order)
                     @php 
-                        $customer = DB::table('customers')->get()
-                            ->where('id', $order->customer_id)
-                            ->first();
-                        $userId = $customer->user_id;
-                        $user = DB::table('users')
-                            ->where('id', $userId)
-                            ->first();
-                        $email = $user->email;
+                        $customer = App\Models\Customer::where('id', $order->customer_id)->firstOrFail();
+                        $email = $customer->user->email;
                     @endphp
     			  
-                    <tr>
+                    <tr class="{{ $order->is_active === 1 ? 'alert-success' : ''}}">
                         <td> {{ $order->id or ''}} </td>
-                        <td> {{ $customer->name . ' ' . $customer->surname}} </td>
+                        <td> <a href="admin/users/edit/{{$customer->user_id}}">{{ $customer->name . ' ' . $customer->surname}} </a></td>
                         <td> {{ $customer->phone or ''}} </td>
                         <td> {{ $email or ''}} </td>
                         <td> {{ $order->address or ''}} </td>
                         <td> {{ $order->order_amount or ''}} </td>
                         <td> <a href="admin/orders/details/{{$order->id}}">Позиции</a></td>
-                        <td> {{ $order->created_at or '' }} </td>
-                        <td> {{ $order->shipdate or '' }} </td>
+                        <td> {{ \Carbon\Carbon::parse($order->created_at)->format('d-m-Y H:i') }} </td>
+                        <td> {{ \Carbon\Carbon::parse($order->shipdate)->format('d-m-Y') }} </td>
                         <td> {{ $order->is_paid ? 'да' : 'нет' }} </td>
                         <td> {{ $order->is_shipped ? 'да' : 'нет'  }} </td>
                         <td> @switch($order->is_active)
@@ -61,7 +56,7 @@
                         @endswitch
                         </td>
                         <td> <a href="admin/orders/edit/{{$order->id}}">Редактировать</a></td>
-                        <td> <a href="admin/orders/delete/{{$order->id}}">Удалить</a></td>
+                        <td> <a href="admin/orders/delete/{{$order->id}}" onclick="return confirm('Вы уверены?');">Удалить</a></td>
                     </tr>
                 @empty
                 	Нет заказов
